requirements.txt
iniconfig==2.3.0
mysql-connector-python==9.5.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
